Fixing contains bug
When clicking exactly on the edge of two fields the game did not
initialize properly. Therefore the init function was updated to keep an
eye on the issue, not setting the indeces. Probably the more important
fix might be the contains function in the Field class, which allowed
some pixellines to be ignored, what could cause some serious trouble.